+++ client/client.go
@@ -1,7 +1,5 @@
-/*
-Package client provides the primary set of types and interfaces needed to implement
-a plugin compatible with keds.
-*/
+// Package client provides the primary set of types and interfaces needed to implement
+// a plugin compatible with keds.
 package client
 
 import (
@@ -21,24 +19,11 @@ import (
 	"google.golang.org/grpc/metadata"
 )
 
-/*
-ClientCallbackHandler implementations should process the main logic of a plugin.
-*/
-type ClientCallbackHandler interface {
-	OnCommandInvoked(client *Client, event *pb.PluginEvent, cmd *cobra.Command, args []string)
-	OnBusEvent(client *Client, event *pb.PluginEvent)
-	OnInitRootCommand(client *Client, cmd *cobra.Command)
-	OnQuit(client *Client)
-	OnRegistered(client *Client)
-	OnConnected(client *Client)
-}
-
-/*
-Client is the primary type that should be used to create plugins. Plugins
-communicate with the plugin host over gRPC.
-*/
+// Client is the primary type that should be used to create plugins. Plugins
+// communicate with the plugin host over gRPC, and use the CallbackHandler to
+// receive plugin lifecycle events.
 type Client struct {
-	CallbackHandler    ClientCallbackHandler
+	callbackHandler    CallbackHandler
 	Conn               *grpc.ClientConn
 	PluginDescriptor   *pb.PluginDescriptor
 	Token              string
@@ -57,8 +42,8 @@ type Client struct {
 	pb.KedsServiceClient
 }
 
-func NewClient(handler ClientCallbackHandler) (client *Client, err error) {
-	client = &Client{CallbackHandler: handler, PluginDescriptor: &pb.PluginDescriptor{}}
+func NewClient(handler CallbackHandler) (client *Client, err error) {
+	client = &Client{callbackHandler: handler, PluginDescriptor: &pb.PluginDescriptor{}}
 	if err = client.loadPluginConfig(); err != nil {
 		return
 	}
@@ -101,7 +86,7 @@ func (c *Client) initRootCommand() {
 			Short: c.PluginDescriptor.GetShortDescription(),
 			Long:  c.PluginDescriptor.GetLongDescription(),
 		}
-		c.CallbackHandler.OnInitRootCommand(c, c.RootCommand)
+		c.callbackHandler.OnInitRootCommand(c, c.RootCommand)
 	}
 }
 
@@ -109,11 +94,11 @@ func (c *Client) Run() (err error) {
 	if err = c.connect(server.EndPoint()); err != nil {
 		return
 	}
-	c.CallbackHandler.OnConnected(c)
+	c.callbackHandler.OnConnected(c)
 	if err = c.register(); err != nil {
 		return
 	}
-	c.CallbackHandler.OnRegistered(c)
+	c.callbackHandler.OnRegistered(c)
 	if c.EventBusStream, err = c.EventBus(c.contextWithToken()); err != nil {
 		return
 	}
@@ -150,7 +135,7 @@ func (c *Client) Quit(code int) {
 	if err := c.EventBusStream.Send(quitEvent); err != nil {
 		log.Printf("failed to send event: %s", err)
 	}
-	c.CallbackHandler.OnQuit(c)
+	c.callbackHandler.OnQuit(c)
 	c.close()
 }
 
@@ -207,11 +192,11 @@ func (c *Client) loop() (err error) {
 }
 
 func (c *Client) onBusEvent(in *pb.PluginEvent) {
-	c.CallbackHandler.OnBusEvent(c, in)
+	c.callbackHandler.OnBusEvent(c, in)
 	if in.GetName() == "keds.command_invoked" && c.RootCommand != nil {
 		c.RootCommand.SetArgs(in.GetArgs())
 		c.RootCommand.Run = func(cmd *cobra.Command, args []string) {
-			c.CallbackHandler.OnCommandInvoked(c, in, cmd, args)
+			c.callbackHandler.OnCommandInvoked(c, in, cmd, args)
 		}
 		if err := c.RootCommand.Execute(); err != nil {
 			log.Printf("error executing command: %s", err)
