+++ plugin/builtin/notifications/console_notifier.go
@@ -1,47 +1,23 @@
 package main
 
 import (
-	"mime"
-	"strings"
-
-	"github.com/talbright/keds/gen/proto"
+	"github.com/talbright/keds/client"
+	"golang.org/x/net/context"
 )
 
-func init() {
-	mime.AddExtensionType(".txt", "text/plain")
+//ConsoleNotifier sends any text/plain messages it receives to the console
+//using the RPC client Printf method.
+type ConsoleNotifier struct{}
+
+//NewConsoleNotifier creates a console notifier that implements the Notifier
+//interface.
+func NewConsoleNotifier(client *client.Client) *ConsoleNotifier {
+	return &ConsoleNotifier{}
 }
 
-type Content struct {
-	Data string
-	Type string
-}
-
-type Message struct {
-	content []Content
-}
-
-func NewMessage(event *proto.PluginEvent) *Message {
-	m := &Message{content: make([]Content, 0)}
-	m.extractContent(event)
-	return m
-}
-
-func (m *Message) GetData(mimeType string) (data string) {
-	for _, v := range m.content {
-		if v.Type == mimeType {
-			return v.Data
-		}
-	}
-	return ""
-}
-
-func (m *Message) extractContent(event *proto.PluginEvent) {
-	for k, v := range event.Data {
-		parts := strings.Split(k, ".")
-		if len(parts) == 2 && parts[0] == "message" {
-			if mimeType := mime.TypeByExtension("." + parts[1]); mimeType != "" {
-				m.content = append(m.content, Content{Data: v, Type: mimeType})
-			}
-		}
+//Notify implementation of Notifier interface
+func (n *ConsoleNotifier) Notify(ctx context.Context, cli *client.Client, notification *Notification) {
+	if data := notification.GetData("text/plain"); data != "" {
+		cli.Printf("[%s] %s", notification.Category, data)
 	}
 }
